Read an .h5ad file and create a SingleCellExperiment.

In file-backed mode, the main X matrix is not read into memory, but references the HDF5 file and its required parts are read on demand. This requires the HDF5Array package to be installed.

Usage

readH5AD(file, backed = FALSE)

Arguments

file

Path to the .h5ad file.

backed

Whether to use file-backed mode.

Value

A SingleCellExperiment.

Examples

data(miniACC, package="MultiAssayExperiment")
writeH5AD(miniACC[[1]], "miniacc.h5ad")
#> Called from: writeH5AD(miniACC[[1]], "miniacc.h5ad")
#> debug: if (is(object, "SingleCellExperiment")) {
#>     write_data_frame(file, "var", rowData(object))
#>     obsm <- reducedDims(object)
#>     if (length(obsm) > 0) {
#>         obsmgrp <- H5Gcreate(file, "obsm")
#>         mapply(function(name, data) {
#>             if (is.data.frame(data)) {
#>                 rownames(data) <- rownames(colData(object))
#>                 write_data_frame(obsmgrp, name, data)
#>             }
#>             else {
#>                 if (length(dim(data)) == 1) 
#>                   data <- as.vector(data)
#>                 else data <- t(data)
#>                 write_matrix(obsmgrp, name, data)
#>             }
#>         }, names(obsm), obsm)
#>         H5Gclose(obsmgrp)
#>     }
#>     lapply(list(list(name = "obsp", names = colPairNames, getter = colPair), 
#>         list(name = "varp", names = rowPairNames, getter = rowPair)), 
#>         function(cp) {
#>             names <- cp$names(object)
#>             if (length(names) > 0) {
#>                 pairgrp <- H5Gcreate(file, cp$name)
#>                 lapply(names, function(cname) {
#>                   write_matrix(pairgrp, cname, cp$getter(object, 
#>                     cname, asSparse = TRUE))
#>                 })
#>                 H5Gclose(pairgrp)
#>             }
#>         })
#>     object <- as(object, "SummarizedExperiment")
#>     written_object <- TRUE
#> }
#> debug: if (is(object, "SummarizedExperiment")) {
#>     write_data_frame(file, "obs", colData(object))
#>     rdata <- rowData(object)
#>     if (ncol(rdata) > 0 || !is.null(rownames(rdata))) 
#>         write_data_frame(file, "var", rdata)
#>     assays <- assays(object)
#>     nassays <- length(assays)
#>     write_matrix(file, "X", assays[[1]])
#>     if (nassays > 1) {
#>         layersgrp <- H5Gcreate(file, "layers")
#>         mapply(function(name, mat) {
#>             write_matrix(layersgrp, name, mat)
#>         }, names(assays[2:nassays]), assays[2:nassays])
#>         H5Gclose(layersgrp)
#>     }
#>     writeList(file, "uns", metadata(object))
#>     if (hasMethod("altExps", class(object))) {
#>         naltexps <- length(altExps(object))
#>         if (naltexps > 1) {
#>             warning("Alternative experiments are currently unsupported. Construct a MultiAssayExperiment object or write them as individual H5AD files.")
#>         }
#>     }
#>     written_object <- TRUE
#> }
#> debug: write_data_frame(file, "obs", colData(object))
#> debug: rdata <- rowData(object)
#> debug: if (ncol(rdata) > 0 || !is.null(rownames(rdata))) write_data_frame(file, 
#>     "var", rdata)
#> debug: write_data_frame(file, "var", rdata)
#> debug: assays <- assays(object)
#> debug: nassays <- length(assays)
#> debug: write_matrix(file, "X", assays[[1]])
#> debug: if (nassays > 1) {
#>     layersgrp <- H5Gcreate(file, "layers")
#>     mapply(function(name, mat) {
#>         write_matrix(layersgrp, name, mat)
#>     }, names(assays[2:nassays]), assays[2:nassays])
#>     H5Gclose(layersgrp)
#> }
#> debug: writeList(file, "uns", metadata(object))
#> debug: if (hasMethod("altExps", class(object))) {
#>     naltexps <- length(altExps(object))
#>     if (naltexps > 1) {
#>         warning("Alternative experiments are currently unsupported. Construct a MultiAssayExperiment object or write them as individual H5AD files.")
#>     }
#> }
#> debug: written_object <- TRUE
#> debug: if (is(object, "Matrix_OR_DelayedMatrix")) {
#>     write_matrix(file, "X", object)
#>     rownames <- rownames(object)
#>     colnames <- colnames(object)
#>     if (is.null(rownames)) 
#>         rownames <- as.character(seq_len(nrow(object)))
#>     if (is.null(colnames)) 
#>         colnames <- as.character(seq_len(ncol(object)))
#>     var <- data.frame(row.names = rownames)
#>     obs <- data.frame(row.names = colnames)
#>     write_data_frame(file, "obs", obs)
#>     write_data_frame(file, "var", var)
#>     written_object <- TRUE
#> }
#> debug: if (!written_object) {
#>     warning("Objects of class ", class(object), " are currently unsupported, skipping...")
#> } else {
#>     write_object_class(file, cls[1])
#>     finalize_anndata_internal(file)
#>     if (need_finalize) 
#>         finalize_anndata(file)
#> }
#> debug: write_object_class(file, cls[1])
#> debug: finalize_anndata_internal(file)
#> debug: if (need_finalize) finalize_anndata(file)
#> debug: finalize_anndata(file)
#> debug: invisible(NULL)
sce <- readH5AD("miniacc.h5ad")